+++ workflows/coarsen_3072_surface_diagnostics/submit_job.sh
@@ -26,7 +26,7 @@ python -m $BEAM_MAIN_MODULE  \
     --region us-central1 \
     --runner DataFlow \
     --setup_file $SETUP_PY \
-    --temp_location gs://vcm-ml-data/tmp_dataflow \
+    --temp_location gs://vcm-ml-scratch/tmp_dataflow \
     --num_workers $NUM_WORKERS \
     --max_num_workers $NUM_WORKERS \
     --disk_size_gb 100 \