+++ workflows/coarsen_restarts/submit_job.sh
@@ -19,7 +19,7 @@ python -m fv3net.pipelines.coarsen_restarts\
     --job_name coarsen-restarts-$user \
     --project vcm-ml \
     --region us-central1 \
-    --temp_location gs://vcm-ml-data/tmp_dataflow \
+    --temp_location gs://vcm-ml-scratch/tmp_dataflow \
     --num_workers 3 \
     --max_num_workers 50 \
     --disk_size_gb 50 \