+++ workflows/prognostic_run_diags/save_prognostic_run_diags.py
@@ -5,6 +5,12 @@ This script computes diagnostics for prognostic runs.
 
 Diagnostics are multiple dimensional curves that can be visualized to give
 a more detailed look into the data underlying the metrics.
+
+Diagnostics are loaded and saved in groupings corresponding to the source of the data
+and its intended use. E.g. the "dycore" grouping includes diagnostics output by the
+dynamical core of the model and saved in `atmos_dt_atmos.tile*.nc` while the "physics"
+grouping contains outputs from the physics routines (`sfc_dt_atmos.tile*.nc` and
+`diags.zarr`).
 """
 import argparse
 import os
@@ -15,11 +21,12 @@ import intake
 import numpy as np
 import xarray as xr
 import shutil
+from dask.diagnostics import ProgressBar
 
 from pathlib import Path
 from toolz import curry
 from collections import defaultdict
-from typing import Tuple, Dict, Callable
+from typing import Tuple, Dict, Callable, Mapping
 
 import vcm
 import load_diagnostic_data as load_diags
@@ -29,15 +36,17 @@ import logging
 logger = logging.getLogger(__file__)
 
 _DIAG_FNS = defaultdict(list)
+_DERIVED_VAR_FNS = defaultdict(Callable)
 
-HORIZONTAL_DIMS = ["grid_xt", "grid_yt", "tile"]
-
+HORIZONTAL_DIMS = ["x", "y", "tile"]
+SECONDS_PER_DAY = 86400
 
 DiagArg = Tuple[xr.Dataset, xr.Dataset, xr.Dataset]
+DiagDict = Mapping[str, xr.DataArray]
 
 
 @curry
-def add_to_diags(diags_key: str, func: Callable[[DiagArg], xr.Dataset]):
+def add_to_diags(diags_key: str, func: Callable[[DiagArg], DiagDict]):
     """
     Add a function to the list of diagnostics to be computed
     for a specified group of data.
@@ -47,10 +56,10 @@ def add_to_diags(diags_key: str, func: Callable[[DiagArg], xr.Dataset]):
         func: a function which computes a set of diagnostics.
             It needs to have the following signature::
 
-                func(prognostic_run_data, verificiation_c48_data, grid)
+                func(prognostic_run_data, verification_c48_data, grid)
 
-            and should return an xarray Dataset of diagnostics.
-            This output will be merged with all other decoratored functions,
+            and should return diagnostics as a dict of xr.DataArrays.
+            This output will be merged with all other decorated functions,
             so some care must be taken to avoid variable and coordinate clashes.
     """
     _DIAG_FNS[diags_key].append(func)
@@ -58,15 +67,16 @@ def add_to_diags(diags_key: str, func: Callable[[DiagArg], xr.Dataset]):
     return func
 
 
-def compute_all_diagnostics(input_datasets: Dict[str, DiagArg]):
+def compute_all_diagnostics(input_datasets: Dict[str, DiagArg]) -> DiagDict:
     """
     Compute all diagnostics for input data.
 
-    Args
-    ----
-    input_datasets:
-        Input datasets with keys corresponding to the appropriate group of
-        diagnostics (_DIAG_FNS) to be run on each data source.
+    Args:
+        input_datasets: Input datasets with keys corresponding to the appropriate group
+        of diagnostics (_DIAG_FNS) to be run on each data source.
+
+    Returns:
+        all computed diagnostics
     """
 
     diags = {}
@@ -85,12 +95,131 @@ def compute_all_diagnostics(input_datasets: Dict[str, DiagArg]):
     return diags
 
 
+@curry
+def add_to_derived_vars(diags_key: str, func: Callable[[xr.Dataset], xr.Dataset]):
+    """Add a function that computes additional derived variables from input datasets.
+    These derived variables should be simple combinations of input variables, not
+    reductions such as global means which are diagnostics to be computed later.
+
+    Args:
+        diags_key: a key for a group of inputs/diagnostics
+        func: a function which adds new variables to a given dataset
+    """
+    _DERIVED_VAR_FNS[diags_key] = func
+
+
+def compute_all_derived_vars(input_datasets: Dict[str, DiagArg]) -> Dict[str, DiagArg]:
+    """Compute derived variables for all input data sources.
+
+    Args:
+        input_datasets: Input datasets with keys corresponding to the appropriate group
+        of inputs/diagnostics.
+
+    Returns:
+        input datasets with derived variables added to prognostic and verification data
+    """
+    for key, func in _DERIVED_VAR_FNS.items():
+        prognostic, verification, grid = input_datasets[key]
+        logger.info(f"Preparing all derived variables for {key} prognostic data")
+        prognostic = prognostic.merge(func(prognostic))
+        logger.info(f"Preparing all derived variables for {key} verification data")
+        verification = verification.merge(func(verification))
+        input_datasets[key] = prognostic, verification, grid
+    return input_datasets
+
+
+@add_to_derived_vars("physics")
+def derived_physics_variables(ds: xr.Dataset) -> xr.Dataset:
+    """Compute derived variables for physics datasets"""
+    arrays = []
+    for func in [
+        _column_pq1,
+        _column_pq2,
+        _column_dq1,
+        _column_dq2,
+        _column_q1,
+        _column_q2,
+    ]:
+        try:
+            arrays.append(func(ds))
+        except (KeyError, AttributeError):  # account for ds[var] and ds.var notations
+            logger.warning(f"Missing variable for calculation in {func.__name__}")
+    return xr.merge(arrays)
+
+
+def _column_pq1(ds: xr.Dataset) -> xr.DataArray:
+    net_heating_arg_labels = [
+        "DLWRFsfc",
+        "DSWRFsfc",
+        "ULWRFsfc",
+        "ULWRFtoa",
+        "USWRFsfc",
+        "USWRFtoa",
+        "DSWRFtoa",
+        "SHTFLsfc",
+        "PRATEsfc",
+    ]
+    net_heating_args = [ds[var] for var in net_heating_arg_labels]
+    column_pq1 = vcm.net_heating(*net_heating_args)
+    column_pq1.attrs = {
+        "long_name": "<pQ1> column integrated heating from physics",
+        "units": "W/m^2",
+    }
+    return column_pq1.rename("column_integrated_pQ1")
+
+
+def _column_pq2(ds: xr.Dataset) -> xr.Dataset:
+    evap = vcm.latent_heat_flux_to_evaporation(ds.LHTFLsfc)
+    column_pq2 = SECONDS_PER_DAY * (evap - ds.PRATEsfc)
+    column_pq2.attrs = {
+        "long_name": "<pQ2> column integrated moistening from physics",
+        "units": "mm/day",
+    }
+    return column_pq2.rename("column_integrated_pQ2")
+
+
+def _column_dq1(ds: xr.Dataset) -> xr.Dataset:
+    column_dq1 = ds.net_heating
+    column_dq1.attrs = {
+        "long_name": "<dQ1> column integrated heating from ML",
+        "units": "W/m^2",
+    }
+    return column_dq1.rename("column_integrated_dQ1")
+
+
+def _column_dq2(ds: xr.Dataset) -> xr.Dataset:
+    column_dq2 = SECONDS_PER_DAY * ds.net_moistening
+    column_dq2.attrs = {
+        "long_name": "<dQ2> column integrated moistening from ML",
+        "units": "mm/day",
+    }
+    return column_dq2.rename("column_integrated_dQ2")
+
+
+def _column_q1(ds: xr.Dataset) -> xr.Dataset:
+    column_q1 = _column_pq1(ds) + _column_dq1(ds)
+    column_q1.attrs = {
+        "long_name": "<Q1> column integrated heating from physics+ML",
+        "units": "W/m^2",
+    }
+    return column_q1.rename("column_integrated_Q1")
+
+
+def _column_q2(ds: xr.Dataset) -> xr.Dataset:
+    column_q2 = _column_pq2(ds) + _column_dq2(ds)
+    column_q2.attrs = {
+        "long_name": "<Q2> column integrated moistening from physics+ML",
+        "units": "mm/day",
+    }
+    return column_q2.rename("column_integrated_Q2")
+
+
 def rms(x, y, w, dims):
     return np.sqrt(((x - y) ** 2 * w).sum(dims) / w.sum(dims))
 
 
 def bias(truth, prediction, w, dims):
-    return ((truth - prediction) * w).sum(dims) / w.sum(dims)
+    return ((prediction - truth) * w).sum(dims) / w.sum(dims)
 
 
 def calc_ds_diurnal_cycle(ds):
@@ -134,9 +263,8 @@ def _add_derived_moisture_diurnal_quantities(ds_run, ds_verif):
         }
     )
 
-    # TODO: add thermo function into top-level import?
-    E_run = vcm.calc.thermo.latent_heat_flux_to_evaporation(ds_run["LHTFLsfc"])
-    E_verif = vcm.calc.thermo.latent_heat_flux_to_evaporation(ds_verif["LHTFLsfc"])
+    E_run = vcm.latent_heat_flux_to_evaporation(ds_run["LHTFLsfc"])
+    E_verif = vcm.latent_heat_flux_to_evaporation(ds_verif["LHTFLsfc"])
     E_diff = E_run - E_verif
     P_diff = ds_run[f"total_P_{filter_flag}"] - ds_verif[f"total_P_{filter_flag}"]
 
@@ -168,7 +296,7 @@ def dump_nc(ds: xr.Dataset, f):
             shutil.copyfileobj(tmp1, f)
 
 
-def _vars_to_diags(suffix, ds, src_ds):
+def _prepare_diag_dict(suffix: str, ds: xr.Dataset, src_ds: xr.Dataset) -> DiagDict:
     # converts dataset variables into diagnostics dict
     # and assigns attrs from src_ds if none are set
 
@@ -186,22 +314,41 @@ def _vars_to_diags(suffix, ds, src_ds):
     return diags
 
 
-@add_to_diags("3H")
+@add_to_diags("dycore")
 def rms_errors(resampled, verification_c48, grid):
     logger.info("Preparing rms errors")
     rms_errors = rms(resampled, verification_c48, grid.area, dims=HORIZONTAL_DIMS)
 
-    return _vars_to_diags("rms_global", rms_errors, resampled)
+    return _prepare_diag_dict("rms_global", rms_errors, resampled)
 
 
-@add_to_diags("3H")
-def global_averages(resampled, verification, grid):
-    logger.info("Preparing global averages")
+@add_to_diags("dycore")
+def global_averages_dycore(resampled, verification, grid):
+    logger.info("Preparing global averages for dycore variables")
     area_averages = (resampled * grid.area).sum(HORIZONTAL_DIMS) / grid.area.sum(
         HORIZONTAL_DIMS
     )
 
-    return _vars_to_diags("global_avg", area_averages, resampled)
+    return _prepare_diag_dict("global_avg", area_averages, resampled)
+
+
+@add_to_diags("physics")
+def global_averages_physics(resampled, verification, grid):
+    logger.info("Preparing global averages for physics variables")
+    area_averages = (resampled * grid.area).sum(HORIZONTAL_DIMS) / grid.area.sum(
+        HORIZONTAL_DIMS
+    )
+
+    return _prepare_diag_dict("global_phys_avg", area_averages, resampled)
+
+
+# TODO: enable this diagnostic once SHiELD physics diags can be loaded efficiently
+# @add_to_diags("physics")
+def global_biases_physics(resampled, verification, grid):
+    logger.info("Preparing global average biases for physics variables")
+    bias_errors = bias(verification, resampled, grid.area, HORIZONTAL_DIMS)
+
+    return _prepare_diag_dict("bias_global_physics", bias_errors, resampled)
 
 
 def diurnal_cycles(resampled, verification, grid):
@@ -220,48 +367,10 @@ def diurnal_cycles(resampled, verification, grid):
     )
 
     # Add to diagnostics
-    verif_diags = _vars_to_diags("verif_diurnal", diurnal_verif, verification)
-    resampled_diags = _vars_to_diags("run_diurnal", diurnal_resampled, resampled)
+    verif_diags = _prepare_diag_dict("verif_diurnal", diurnal_verif, verification)
+    resampled_diags = _prepare_diag_dict("run_diurnal", diurnal_resampled, resampled)
 
     return dict(**verif_diags, **resampled_diags)
-
-
-def open_tiles(path):
-    return xr.open_mfdataset(path + ".tile?.nc", concat_dim="tile", combine="nested")
-
-
-# TODO: This operation has mostly been set up in load_diagnostic_data, still need
-#       to figure out where resampling to 3H makes sense (a transform) and how to
-#       handle multi-timescale verification sources
-def load_data_3H(url, grid_spec, catalog):
-    logger.info(f"Processing 3H data from run directory at {url}")
-
-    # open grid
-    logger.info("Opening Grid Spec")
-    grid_c384 = open_tiles(grid_spec)
-
-    # open verification
-    logger.info("Opening verification data")
-    verification = catalog["40day_c384_atmos_8xdaily"].to_dask()
-    verification = verification.merge(grid_c384)
-
-    # block average data
-    logger.info("Block averaging the verification data")
-    verification_c48 = vcm.cubedsphere.weighted_block_average(
-        verification, verification.area, 8, x_dim="grid_xt", y_dim="grid_yt"
-    )
-
-    # open data
-    atmos_diag_url = os.path.join(url, "atmos_dt_atmos")
-    logger.info(f"Opening diagnostic data at {atmos_diag_url}")
-    ds = open_tiles(atmos_diag_url).load()
-    resampled = ds.resample(time="3H", label="right").nearest()
-
-    verification_c48 = verification_c48.sel(
-        time=resampled.time[:-1]
-    )  # don't use last time point. there is some trouble
-
-    return resampled, verification_c48, verification_c48[["area"]]
 
 
 def _catalog():
@@ -289,31 +398,29 @@ if __name__ == "__main__":
 
     catalog = intake.open_catalog(args.catalog)
     input_data = {}
-    resampled, verification, grid = load_data_3H(args.url, args.grid_spec, catalog)
-    input_data["3H"] = (resampled, verification, grid)
+    input_data["dycore"] = load_diags.load_dycore(args.url, args.grid_spec, catalog)
+    input_data["physics"] = load_diags.load_physics(args.url, args.grid_spec, catalog)
 
-    # Data loaded at original time resolution
-    # verif_base = load_diags.load_verification(
-    #     ["40day_c384_diags_time_avg"], catalog, 8, grid.area
-    # )
-    # data_base = load_diags.load_diagnostics(args.url)
+    # add derived variables
+    input_data = compute_all_derived_vars(input_data)
 
     # begin constructing diags
     diags = {}
 
     # maps
-    diags["pwat_run_initial"] = resampled.PWAT.isel(time=0)
-    diags["pwat_run_final"] = resampled.PWAT.isel(time=-2)
-    diags["pwat_verification_final"] = verification.PWAT.isel(time=-2)
+    diags["pwat_run_initial"] = input_data["dycore"][0].PWAT.isel(time=0)
+    diags["pwat_run_final"] = input_data["dycore"][0].PWAT.isel(time=-2)
+    diags["pwat_verification_final"] = input_data["dycore"][0].PWAT.isel(time=-2)
 
     diags.update(compute_all_diagnostics(input_data))
 
     # add grid vars
     diags = xr.Dataset(diags, attrs=attrs)
-    diags = diags.merge(grid)
+    diags = diags.merge(input_data["dycore"][2])
 
     logger.info("Forcing computation.")
-    diags = diags.load()
+    with ProgressBar():
+        diags = diags.load()
 
     logger.info(f"Saving data to {args.output}")
     diags.to_netcdf(args.output)