+++ HISTORY.rst
@@ -24,6 +24,9 @@ training data from multiple sources into a set of diagnostic variables
 * Add optional arg `--timesteps-file` to fv3net.regression.sklearn to provide a list of timesteps to use 
  in conjunction with the config arg `num_timesteps_per_batch`. The training config arg `num_batches` is now
  deprecated in favor of providing a list of timesteps.
+* Remove usage of "tmp_dataflow" directory from `vcm.cloud.gcs` testing infrastructure and skip extract tests
+* Update prognostic run report to include global averages of physics variables and heating/moistening
+* Prognostic report requires that prognostic runs have been post-processed before being passed to the report workflow
 
 
 0.2.1 (2020-05-15)